requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.0
psycopg2-binary>=2.9
ollama>=0.4
langchain-core>=0.3
langchain-ollama>=0.3